End type, close primitive and some tests
Implemented the End type, which is a linear session type such that
End;T is equivalent to End for all T (see issue #138).
The close primitive consumes a channel of this type.
Included some unit tests, along with program tests based on #138.
Refactored some tests that used End as a label.