Arrow subtyping, tests and tuning.
Implements function-type subtyping suggested in a comment in #132 and
further commented on in #127. Now unrestricted functions can be used
wherever a linear one is needed (since unrestricted resources should be
able to be used linearly.)
Added two program tests to validate this new feature.
Adjusted some functions in the prelude to the new type of fork, and
tuned existing program tests accordingly.